Fixed endpoints with some id validations between service/provider

+++ src/service/service.service.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { BusinessService } from '../business/business.service';
 import { Repository } from 'typeorm';
 import { CreateServiceDto } from './dto/create-service.dto';
 import { UpadateServiceDto } from './dto/update-service.dto';
 import { Service } from './service.entity';
+import { ProviderEntity } from '../provider/provider.entity';
 
 @Injectable()
 export class ServiceService {
@@ -14,10 +15,19 @@ export class ServiceService {
     private readonly businessService: BusinessService,
   ) {}
 
-  async createService(serviceData: CreateServiceDto) {
+  async createService(serviceData: CreateServiceDto, provider: ProviderEntity) {
     const business = await this.businessService.getBusinessById(
       serviceData.businessId,
     );
+
+    if (business.provider.id !== provider.id) {
+      throw new HttpException(
+        {
+          message: "The id's dont match.",
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
 
     const service = this.serviceRepository.create({
       ...serviceData,
@@ -41,12 +51,25 @@ export class ServiceService {
     const service = await this.serviceRepository
       .createQueryBuilder('service')
       .where({ id: id })
+      .leftJoinAndSelect('service.business', 'business')
+      .leftJoinAndSelect('business.provider', 'provider')
       .getOne();
 
     return service;
   }
 
-  async deleteServiceById(id: number) {
+  async deleteServiceById(id: number, provider: ProviderEntity) {
+    const service = await this.getServicesById(id);
+
+    if (service.business.provider.id !== provider.id && !service) {
+      throw new HttpException(
+        {
+          message: "The id's dont match.",
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     await this.serviceRepository
       .createQueryBuilder('service')
       .delete()
@@ -56,7 +79,21 @@ export class ServiceService {
     return 'service deleted';
   }
 
-  async updateService(id: number, UpdateServiceBody: UpadateServiceDto) {
+  async updateService(
+    id: number,
+    UpdateServiceBody: UpadateServiceDto,
+    provider: ProviderEntity,
+  ) {
+    const service = await this.getServicesById(id);
+
+    if (service.business.provider.id !== provider.id && !service) {
+      throw new HttpException(
+        {
+          message: "The id's dont match.",
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
     await this.serviceRepository.update(id, UpdateServiceBody);
 
     return await this.serviceRepository.findOne(id);
